Fix leaderboard
At first, the `getLeaderboard` function was returning and incorrect amount of players. Turns out, the issue was that I was fetching players by `initTimestamp_gt` and just feeding the last timestamp returned by that into the next query. The problem is, the last timestamp isn't necessarily the biggest one! So the solution was to include the `orderBy` option to make sure timestamp were sorted in sequential order